refactor(Game): simply code
