+++ LL3D.sln
@@ -21,16 +21,24 @@ Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x64 = Debug|x64
 		Debug|x86 = Debug|x86
+		MinSizeRel|x64 = MinSizeRel|x64
+		MinSizeRel|x86 = MinSizeRel|x86
 		Profile|x64 = Profile|x64
 		Profile|x86 = Profile|x86
 		Release|x64 = Release|x64
 		Release|x86 = Release|x86
+		RelWithDebInfo|x64 = RelWithDebInfo|x64
+		RelWithDebInfo|x86 = RelWithDebInfo|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{543B5569-7C16-41AF-95A2-62E39C0576BD}.Debug|x64.ActiveCfg = Debug|x64
 		{543B5569-7C16-41AF-95A2-62E39C0576BD}.Debug|x64.Build.0 = Debug|x64
 		{543B5569-7C16-41AF-95A2-62E39C0576BD}.Debug|x86.ActiveCfg = Debug|Win32
 		{543B5569-7C16-41AF-95A2-62E39C0576BD}.Debug|x86.Build.0 = Debug|Win32
+		{543B5569-7C16-41AF-95A2-62E39C0576BD}.MinSizeRel|x64.ActiveCfg = Release|x64
+		{543B5569-7C16-41AF-95A2-62E39C0576BD}.MinSizeRel|x64.Build.0 = Release|x64
+		{543B5569-7C16-41AF-95A2-62E39C0576BD}.MinSizeRel|x86.ActiveCfg = Release|Win32
+		{543B5569-7C16-41AF-95A2-62E39C0576BD}.MinSizeRel|x86.Build.0 = Release|Win32
 		{543B5569-7C16-41AF-95A2-62E39C0576BD}.Profile|x64.ActiveCfg = Release|x64
 		{543B5569-7C16-41AF-95A2-62E39C0576BD}.Profile|x64.Build.0 = Release|x64
 		{543B5569-7C16-41AF-95A2-62E39C0576BD}.Profile|x86.ActiveCfg = Release|Win32
@@ -39,10 +47,18 @@ Global
 		{543B5569-7C16-41AF-95A2-62E39C0576BD}.Release|x64.Build.0 = Release|x64
 		{543B5569-7C16-41AF-95A2-62E39C0576BD}.Release|x86.ActiveCfg = Release|Win32
 		{543B5569-7C16-41AF-95A2-62E39C0576BD}.Release|x86.Build.0 = Release|Win32
+		{543B5569-7C16-41AF-95A2-62E39C0576BD}.RelWithDebInfo|x64.ActiveCfg = Release|x64
+		{543B5569-7C16-41AF-95A2-62E39C0576BD}.RelWithDebInfo|x64.Build.0 = Release|x64
+		{543B5569-7C16-41AF-95A2-62E39C0576BD}.RelWithDebInfo|x86.ActiveCfg = Release|Win32
+		{543B5569-7C16-41AF-95A2-62E39C0576BD}.RelWithDebInfo|x86.Build.0 = Release|Win32
 		{7AA4C43B-094F-4F3C-AA48-6ABE085C564A}.Debug|x64.ActiveCfg = Debug|x64
 		{7AA4C43B-094F-4F3C-AA48-6ABE085C564A}.Debug|x64.Build.0 = Debug|x64
 		{7AA4C43B-094F-4F3C-AA48-6ABE085C564A}.Debug|x86.ActiveCfg = Debug|Win32
 		{7AA4C43B-094F-4F3C-AA48-6ABE085C564A}.Debug|x86.Build.0 = Debug|Win32
+		{7AA4C43B-094F-4F3C-AA48-6ABE085C564A}.MinSizeRel|x64.ActiveCfg = Release|x64
+		{7AA4C43B-094F-4F3C-AA48-6ABE085C564A}.MinSizeRel|x64.Build.0 = Release|x64
+		{7AA4C43B-094F-4F3C-AA48-6ABE085C564A}.MinSizeRel|x86.ActiveCfg = Release|Win32
+		{7AA4C43B-094F-4F3C-AA48-6ABE085C564A}.MinSizeRel|x86.Build.0 = Release|Win32
 		{7AA4C43B-094F-4F3C-AA48-6ABE085C564A}.Profile|x64.ActiveCfg = Release|x64
 		{7AA4C43B-094F-4F3C-AA48-6ABE085C564A}.Profile|x64.Build.0 = Release|x64
 		{7AA4C43B-094F-4F3C-AA48-6ABE085C564A}.Profile|x86.ActiveCfg = Release|Win32
@@ -51,10 +67,18 @@ Global
 		{7AA4C43B-094F-4F3C-AA48-6ABE085C564A}.Release|x64.Build.0 = Release|x64
 		{7AA4C43B-094F-4F3C-AA48-6ABE085C564A}.Release|x86.ActiveCfg = Release|Win32
 		{7AA4C43B-094F-4F3C-AA48-6ABE085C564A}.Release|x86.Build.0 = Release|Win32
+		{7AA4C43B-094F-4F3C-AA48-6ABE085C564A}.RelWithDebInfo|x64.ActiveCfg = Release|x64
+		{7AA4C43B-094F-4F3C-AA48-6ABE085C564A}.RelWithDebInfo|x64.Build.0 = Release|x64
+		{7AA4C43B-094F-4F3C-AA48-6ABE085C564A}.RelWithDebInfo|x86.ActiveCfg = Release|Win32
+		{7AA4C43B-094F-4F3C-AA48-6ABE085C564A}.RelWithDebInfo|x86.Build.0 = Release|Win32
 		{DF460EAB-570D-4B50-9089-2E2FC801BF38}.Debug|x64.ActiveCfg = Debug|x64
 		{DF460EAB-570D-4B50-9089-2E2FC801BF38}.Debug|x64.Build.0 = Debug|x64
 		{DF460EAB-570D-4B50-9089-2E2FC801BF38}.Debug|x86.ActiveCfg = Debug|Win32
 		{DF460EAB-570D-4B50-9089-2E2FC801BF38}.Debug|x86.Build.0 = Debug|Win32
+		{DF460EAB-570D-4B50-9089-2E2FC801BF38}.MinSizeRel|x64.ActiveCfg = Release|x64
+		{DF460EAB-570D-4B50-9089-2E2FC801BF38}.MinSizeRel|x64.Build.0 = Release|x64
+		{DF460EAB-570D-4B50-9089-2E2FC801BF38}.MinSizeRel|x86.ActiveCfg = Release|Win32
+		{DF460EAB-570D-4B50-9089-2E2FC801BF38}.MinSizeRel|x86.Build.0 = Release|Win32
 		{DF460EAB-570D-4B50-9089-2E2FC801BF38}.Profile|x64.ActiveCfg = Profile|x64
 		{DF460EAB-570D-4B50-9089-2E2FC801BF38}.Profile|x64.Build.0 = Profile|x64
 		{DF460EAB-570D-4B50-9089-2E2FC801BF38}.Profile|x86.ActiveCfg = Profile|Win32
@@ -63,10 +87,18 @@ Global
 		{DF460EAB-570D-4B50-9089-2E2FC801BF38}.Release|x64.Build.0 = Release|x64
 		{DF460EAB-570D-4B50-9089-2E2FC801BF38}.Release|x86.ActiveCfg = Release|Win32
 		{DF460EAB-570D-4B50-9089-2E2FC801BF38}.Release|x86.Build.0 = Release|Win32
+		{DF460EAB-570D-4B50-9089-2E2FC801BF38}.RelWithDebInfo|x64.ActiveCfg = Release|x64
+		{DF460EAB-570D-4B50-9089-2E2FC801BF38}.RelWithDebInfo|x64.Build.0 = Release|x64
+		{DF460EAB-570D-4B50-9089-2E2FC801BF38}.RelWithDebInfo|x86.ActiveCfg = Release|Win32
+		{DF460EAB-570D-4B50-9089-2E2FC801BF38}.RelWithDebInfo|x86.Build.0 = Release|Win32
 		{A97B0068-D04A-44EA-BC9A-8F9AB81FEFB6}.Debug|x64.ActiveCfg = Debug|x64
 		{A97B0068-D04A-44EA-BC9A-8F9AB81FEFB6}.Debug|x64.Build.0 = Debug|x64
 		{A97B0068-D04A-44EA-BC9A-8F9AB81FEFB6}.Debug|x86.ActiveCfg = Debug|Win32
 		{A97B0068-D04A-44EA-BC9A-8F9AB81FEFB6}.Debug|x86.Build.0 = Debug|Win32
+		{A97B0068-D04A-44EA-BC9A-8F9AB81FEFB6}.MinSizeRel|x64.ActiveCfg = Release|x64
+		{A97B0068-D04A-44EA-BC9A-8F9AB81FEFB6}.MinSizeRel|x64.Build.0 = Release|x64
+		{A97B0068-D04A-44EA-BC9A-8F9AB81FEFB6}.MinSizeRel|x86.ActiveCfg = Release|Win32
+		{A97B0068-D04A-44EA-BC9A-8F9AB81FEFB6}.MinSizeRel|x86.Build.0 = Release|Win32
 		{A97B0068-D04A-44EA-BC9A-8F9AB81FEFB6}.Profile|x64.ActiveCfg = Release|x64
 		{A97B0068-D04A-44EA-BC9A-8F9AB81FEFB6}.Profile|x64.Build.0 = Release|x64
 		{A97B0068-D04A-44EA-BC9A-8F9AB81FEFB6}.Profile|x86.ActiveCfg = Release|Win32
@@ -75,6 +107,10 @@ Global
 		{A97B0068-D04A-44EA-BC9A-8F9AB81FEFB6}.Release|x64.Build.0 = Release|x64
 		{A97B0068-D04A-44EA-BC9A-8F9AB81FEFB6}.Release|x86.ActiveCfg = Release|Win32
 		{A97B0068-D04A-44EA-BC9A-8F9AB81FEFB6}.Release|x86.Build.0 = Release|Win32
+		{A97B0068-D04A-44EA-BC9A-8F9AB81FEFB6}.RelWithDebInfo|x64.ActiveCfg = Release|x64
+		{A97B0068-D04A-44EA-BC9A-8F9AB81FEFB6}.RelWithDebInfo|x64.Build.0 = Release|x64
+		{A97B0068-D04A-44EA-BC9A-8F9AB81FEFB6}.RelWithDebInfo|x86.ActiveCfg = Release|Win32
+		{A97B0068-D04A-44EA-BC9A-8F9AB81FEFB6}.RelWithDebInfo|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
