+++ Source/Editor/Editor.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <Windows.h>
-#include "Application.h"
+#include "Game.h"
 #include "EditorCamera.h"
 #include "FirstPersonalCamera.h"
 #include "Engine.h"
@@ -12,9 +12,6 @@ using namespace DirectX;
 class Editor: public LL3D::Game {
 public:
   Editor();
-
-  // fps <= 0 means unlimited.
-  void SetMaxFPS(int fps);
 
 protected:
   virtual void Update() override;
@@ -29,15 +26,11 @@ protected:
   virtual void OnResize();
 
 private:
-  void DoUpdate();
-
   EditorCamera editor_camera_;
   LL3D::FirstPersonalCamera first_personal_camera_;
   LL3D::Engine engine_;
   LL3D::Stopwatch timer_;
   LL3D::Lights lights_;
   
-  int td_before_last_draw_ = 0;
-  int max_fps_ = 0;
   LL3D::IntPoint2 last_mouse_position_;
 };
